refactor(create_raw.py): Renaming from vis_create_raw to create_raw
