+++ setup.py
@@ -18,8 +18,8 @@ setup(
             'nss_simulator=vipersci.carto.nss_simulator:main',
             'tri2gpkg=vipersci.carto.tri2gpkg:main',
             "template_test=vipersci.pds.template_test:main",
-            "vis_create_raw=vipersci.pds.vis_create_raw:main",
-            "vis_create_raw_tif=vipersci.pds.vis_create_raw_tif:main",
+            "vis_create_raw=vipersci.vis.pds.create_raw:main",
+            "vis_create_raw_tif=vipersci.vis.pds.create_raw_tif:main",
         ],
     },
     install_requires=requirements,
